+++ packages/components-js/projects/partials/scripts/generateInitialStylesPartial.ts
@@ -18,38 +18,27 @@ import {
 
 // TODO: remove skeleton styles after all are hydrated
 
-const skeletonTagNamesTypeLiteral = joinArrayElementsToString(SKELETON_TAG_NAMES, ' | ');
-
 const tagNames = joinArrayElementsToString(TAG_NAMES.filter((x) => !INTERNAL_TAG_NAMES.includes(x)));
-
 const tagNamesWithSkeleton = joinArrayElementsToString(SKELETON_TAG_NAMES);
 
 // includes skeleton styles when SKELETONS_ACTIVE is set to true
 export const generateInitialStylesPartial = (): string => {
-  // 'any' is fallback when SKELETON_TAG_NAMES is an empty array because shared wasn't built yet
-  const types = `${SKELETONS_ACTIVE ? `export type SkeletonTagName = ${skeletonTagNamesTypeLiteral || 'any'};` : ''}
-
-  type GetInitialStylesOptions = {
-  ${SKELETONS_ACTIVE ? 'skeletonTagNames?: SkeletonTagName[];' : ''}
+  const types = `type GetInitialStylesOptions = {${SKELETONS_ACTIVE ? '\n  skeletonTagNames?: SkeletonTagName[];' : ''}
   prefix?: string;
   ${withoutTagsOption}
   format?: Format;
 };
-type GetInitialStylesOptionsFormatHtml = Omit<GetInitialStylesOptions, 'withoutTags'> & {
-  format: 'html';
-};
-type GetInitialStylesOptionsFormatJsx = Omit<GetInitialStylesOptions, 'withoutTags'> & {
-   format: 'jsx';
-};
+type GetInitialStylesOptionsFormatHtml = Omit<GetInitialStylesOptions, 'withoutTags'> & { format: 'html' };
+type GetInitialStylesOptionsFormatJsx = Omit<GetInitialStylesOptions, 'withoutTags'> & { format: 'jsx' };
 type GetInitialStylesOptionsWithoutTags = Omit<GetInitialStylesOptions, 'format'>;`;
 
-  const skeletonTypes = SKELETONS_ACTIVE
-    ? `type SkeletonStylesOptions = {
+  const skeletonTypes = `type SkeletonTagName = ${joinArrayElementsToString(SKELETON_TAG_NAMES, ' | ')};
+
+type GetSkeletonStylesOptions = {
   prefixedTagNamesWithSkeleton: string[];
   prefixedUnusedTagNamesWithSkeleton: string[];
   prefix?: string;
-}`
-    : '';
+};`;
 
   const skeletonKeyframes = '@keyframes opacity{0%{opacity:0.35}50%{opacity:0.15}100%{opacity:0.35}}';
 
@@ -72,12 +61,13 @@ type GetInitialStylesOptionsWithoutTags = Omit<GetInitialStylesOptions, 'format'
 export function getInitialStyles(opts?: GetInitialStylesOptionsFormatJsx): JSX.Element;
 export function getInitialStyles(opts?: GetInitialStylesOptionsWithoutTags): string;
 export function getInitialStyles(opts?: GetInitialStylesOptions): string | JSX.Element {
-  const { ${SKELETONS_ACTIVE ? 'skeletonTagNames, ' : ''}prefix, withoutTags, format }: GetInitialStylesOptions = {
-    ${SKELETONS_ACTIVE ? 'skeletonTagNames: [],' : ''}
+  const { ${SKELETONS_ACTIVE ? 'skeletonTagNames, ' : ''}prefix, withoutTags, format }: GetInitialStylesOptions = {${
+    SKELETONS_ACTIVE ? '\n    skeletonTagNames: [],' : ''
+  }
     prefix: '',
     withoutTags: false,
     format: 'html',
-    ...opts
+    ...opts,
   };
 
   const tagNames = [${tagNames}];
@@ -98,44 +88,49 @@ Please use only valid component tag names:
 
   const usedTagNamesWithSkeleton = tagNamesWithSkeleton.filter((skeletonTagName) => skeletonTagNames.includes(skeletonTagName));
   const prefixedTagNamesWithSkeleton = getPrefixedTagNames(usedTagNamesWithSkeleton, prefix);
-  const prefixedUnusedTagNamesWithSkeleton = getPrefixedTagNames(tagNamesWithSkeleton.filter((skeletonTagName) => !skeletonTagNames.includes(skeletonTagName)), prefix);`
+  const prefixedUnusedTagNamesWithSkeleton = getPrefixedTagNames(tagNamesWithSkeleton.filter((skeletonTagName) => !skeletonTagNames.includes(skeletonTagName)), prefix);\n`
       : ''
   }
-  const usesSkeleton = ${SKELETONS_ACTIVE ? `usedTagNamesWithSkeleton.length ? ' uses-skeleton="true"': ''` : `''`};
-  const usesSkeletonJsx = ${
-    SKELETONS_ACTIVE ? `usedTagNamesWithSkeleton.length ? {"uses-skeleton": 'true'} : {}` : undefined
-  };
-  const initialVisibilityHiddenStyles = prefixedTagNames.join(',') + '{visibility:hidden}';
+  throwIfRunInBrowser('getInitialStyles');
 
-  const mergedStyles = \`\${initialVisibilityHiddenStyles}${
+  const styleAttributes = ['pds-initial-styles'${
+    SKELETONS_ACTIVE ? `, usedTagNamesWithSkeleton.length && 'uses-skeleton'` : ''
+  }].filter(x => x).join(' ');
+  const styleProps = { 'pds-initial-styles': 'true'${
+    SKELETONS_ACTIVE ? `, ...(usedTagNamesWithSkeleton.length && { 'uses-skeleton': 'true' })` : ''
+  } };
+
+  const initialStyles = prefixedTagNames.join(',') + '{visibility:hidden}.hydrated{visibility:inherit}';
+  const styles = initialStyles${
     SKELETONS_ACTIVE
-      ? '${getSkeletonStyles({prefixedTagNamesWithSkeleton,prefixedUnusedTagNamesWithSkeleton, prefix})}'
+      ? '\n    + getSkeletonStyles({prefixedTagNamesWithSkeleton,prefixedUnusedTagNamesWithSkeleton, prefix})'
       : ''
-  }\`;
-  const markup = format === 'html' ?  \`<style\$\{usesSkeleton\}>\${mergedStyles}</style>\` : <style dangerouslySetInnerHTML={{__html: mergedStyles}} {...usesSkeletonJsx}/>;
+  };
+
+  const markup = format === 'html'
+    ? \`<style \$\{styleAttributes\}>\${styles}</style>\`
+    : <style  {...styleProps} dangerouslySetInnerHTML={{ __html: styles }} />;
 
   return withoutTags
-    ? mergedStyles
+    ? styles
     : markup;
-};`;
+}`;
 
-  const skeletonStylesFunction = SKELETONS_ACTIVE
-    ? `const getSkeletonStyles = (opts?: SkeletonStylesOptions): string => {
-  const options: SkeletonStylesOptions = {
-     prefixedTagNamesWithSkeleton: [],
-     prefixedUnusedTagNamesWithSkeleton: [],
-     prefix: '',
-     ...opts
+  const skeletonStylesFunction = `const getSkeletonStyles = (opts?: GetSkeletonStylesOptions): string => {
+  const { prefixedTagNamesWithSkeleton, prefixedUnusedTagNamesWithSkeleton, prefix }: GetSkeletonStylesOptions = {
+    prefixedTagNamesWithSkeleton: [],
+    prefixedUnusedTagNamesWithSkeleton: [],
+    prefix: '',
+    ...opts,
   };
-  const { prefixedTagNamesWithSkeleton, prefixedUnusedTagNamesWithSkeleton, prefix } = options;
 
   const skeletonStylesWithKey = ${JSON.stringify(minifiedSkeletonStyles)};
-  let skeletonStyles = prefixedTagNamesWithSkeleton.map((prefixedTagName)=>{
+  let skeletonStyles = prefixedTagNamesWithSkeleton.map((prefixedTagName) => {
     let tagNameToFind = prefixedTagName;
 
     // if prefix is used it has to be removed and the tagName has to be reassigned
     // in order to find tagName in keys of skeletonStyles
-    if(prefix){
+    if (prefix) {
       const prefixRegExp = new RegExp(\`\${prefix}-\`, 'g');
       tagNameToFind = prefixedTagName.replace(prefixRegExp, '');
     }
@@ -154,7 +149,7 @@ Please use only valid component tag names:
         skeletonStyleKey.split('|').forEach(key => {
           skeletonStyle = skeletonStyle.replace(new RegExp(\`(\${key}(?!-))\`, 'g'), prefix + '-' + key);
         });
-      };
+      }
 
       // if tagName is found, the key-value-pair can be removed since the style is already applied
       // e.g. 'p-button' is found => check for 'p-link' or 'p-link-social' is redundant
@@ -175,13 +170,16 @@ Please use only valid component tag names:
 
   const result = skeletonStyles + \`\${prefixedTagNamesWithSkeleton.length ? \`\${skeletonStylesWithKey.pseudo}${skeletonKeyframes}\` : ''}\`;
   return result;
-};`
-    : '';
+};`;
 
   const helperFunction = `const getPrefixedTagNames = (tagNames: string[], prefix?: string): string[] => {
   return prefix ? tagNames.map((x) => \`\${prefix}-\${x}\`) : tagNames;
-}
-`;
+};`;
 
-  return [types, skeletonTypes, helperFunction, initialStylesFunction, skeletonStylesFunction].join('\n\n');
+  return [
+    types,
+    initialStylesFunction,
+    helperFunction,
+    ...(SKELETONS_ACTIVE ? [skeletonTypes, skeletonStylesFunction] : []),
+  ].join('\n\n');
 };