+++ packages/components-react/src/pages/ToastBasic.tsx
@@ -19,7 +19,7 @@ export const ToastBasicPage = (): JSX.Element => {
 
   return (
     <>
-      <style children={style} />
+      <style dangerouslySetInnerHTML={{ __html: style }} />
 
       <div className="playground light" title="should render toast neutral on light background">
         <PToast />