+++ packages/storefront/tests/e2e/helpers/puppeteer-helper.ts
@@ -1,9 +1,9 @@
 import { ConsoleMessage, Page } from 'puppeteer';
 
-let consoleMessages: ConsoleMessage[] = [];
+const consoleMessages: ConsoleMessage[] = [];
 
 export const initConsoleObserver = (page: Page): void => {
-  consoleMessages = []; // reset
+  consoleMessages.length = 0; // reset
 
   page.on('console', (msg) => {
     if (msg.type() === 'error') {