+++ packages/utilities/projects/utilities/src/js/font/index.ts
@@ -1,8 +1,9 @@
 import { fontBehavior } from './font-behavior';
 import { fontFamily } from './font-family';
+import { fontHyphenation } from './font-hyphenation';
 import { fontSize } from './font-size';
 import { fontWeight } from './font-weight';
 import { fontStyle } from './font-style';
 import { fontVariant } from './font-variant';
 
-export { fontBehavior, fontFamily, fontSize, fontWeight, fontStyle, fontVariant };
+export { fontBehavior, fontFamily, fontHyphenation, fontSize, fontWeight, fontStyle, fontVariant };